requirements.txt
flask
pymongo
openai
python-dotenv
certifi
dnspython
requests
pydantic
